Move source generator validation to AnalyzerTest
