+++ tests/Microsoft.CodeAnalysis.Testing/Microsoft.CodeAnalysis.Testing.Utilities/TestGenerators/AddEmptyFile.cs
